Changes zPosition of control pad to allow fireballs to go behind control pad

+++ GameElements/Fireball.swift
@@ -18,7 +18,7 @@ class Fireball: SKSpriteNode {
         super.init(texture: texture, color: color, size: size)
         setupFireball()
         setupPhysicsBody()
-        zPosition = CGFloat(GameScene.Layer.Game.rawValue)
+        zPosition = GameScene.Layer.Game.rawValue
     }
 
     required init?(coder aDecoder: NSCoder) {
